+++ backend/src/shops/errors/index.ts
@@ -1,1 +1,2 @@
-export * from './ShopAlreadyExistsError';
+export * from "./ShopAlreadyExistsError";
+export * from "./ShopNotFoundError";
